requirements.txt
torchvision==0.8.1
torch==1.7.0
pandas==1.1.3
requests==2.24.0
numpy==1.19.3
opencv_python==4.4.0.44
Flask==1.1.2
tqdm==4.51.0
Pillow==8.0.1
